Add path log metadata to top of parse result.
When a Tramp record is created by make-tramp, the grammar tree is
walked and the get-in style path is attached as metadata to each node
in the grammar tree. Also a new :path-log atom is added to the Tramp
record.

In push-result (during parsing) the current path is retrieved from the
metadata of node in the grammar tree being processed. The path is then
added to the :path-log of the tramp object.

When the final result is returned, the :path-log from the tramp object
is added to the metadata of the top of the result returned.