Fix path-log order and repeated starting tokens.
The following grammar has alternations that both start with the same
initial token:
    foo = '0'
        | '0' '1' ;

This was causing problems because when instaparse was looking up the
'0' node, there was only one instance of it and the path information
was being stored in metadata which mean that the metadata was for the
path of the second alteration even when the match was for the first
alteration. By moving the path information to a field of the parse node
itself this causes the nodes to not be unique during lookup.

In addition, the ordering of the path-log elements returned was wrong
because we stored the path-log on the top-level Tramp element which
meant the ordering was the stack-based processing order rather than
the actual order that the path elements were encountered. This moves
tracking of path information from the Tramp to metadata on result
nodes so that we can build up the actual parse order.